moved helpers out of rust oacle
